requirements.txt
streamlit
pandas
pillow
qrcode
openpyxl